Move rowSorter() to src/components/ui/table/utils.ts
